clear /lib

+++ mirai-console-wrapper/src/main/kotlin/net/mamoe/mirai/console/wrapper/LibManager.kt
@@ -20,6 +20,9 @@ object LibManager{
     val dynamic = false;
 
     fun clearLibs(){
+        if(!dynamic){
+            return;
+        }
         libPath.listFiles()?.forEach {
             it.delete()
         }
